employee_management_system/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
mysql-connector-python==8.1.0
marshmallow==3.21.1
